Hold List of investigations in one registration

+++ labsync-investigations-microservice/src/main/java/com/labsyncplus/labsync_investigations_microservice/dao/InvestigationDao.java
@@ -3,5 +3,5 @@ package com.labsyncplus.labsync_investigations_microservice.dao;
 import com.labsyncplus.labsync_investigations_microservice.model.entity.Investigation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InvestigationDao extends JpaRepository<Investigation, Integer> {
+public interface InvestigationDao extends JpaRepository<Investigation, Long> {
 }
